DEX-724 One WebSocket connection
Review fixes
