+++ dex/src/main/scala/com/wavesplatform/dex/api/websockets/package.scala
@@ -1,6 +1,5 @@
 package com.wavesplatform.dex.api
 
-import com.wavesplatform.dex.domain.asset.AssetPair
 import play.api.libs.json._
 
 package object websockets {
@@ -16,15 +15,5 @@ package object websockets {
       case x => JsError(JsPath, s"Can't read Double from ${x.getClass.getName}")
     },
     tjs = Writes(x => JsString(x.toString))
-  )
-
-  val assetPairKeyAsStringFormat: Format[AssetPair] = Format(
-    fjs = Reads {
-      case JsString(x) => AssetPair.extractAssetPair(x).fold(e => JsError(e.getMessage), JsSuccess(_))
-      case x           => JsError(JsPath, s"Expected a string, but got ${x.toString().take(10)}...")
-    },
-    tjs = Writes { x =>
-      JsString(x.key)
-    }
   )
 }
